# backend/api/openapi.yaml
openapi: "3.1.0"
info:
    title: "Roshnii API"
    version: "0.1.0" # Updated version
    description: "API for Roshnii - Photo storage and management service."
servers:
    - url: "http://localhost:8080/api" # Base path for API endpoints
      description: "Development server"
paths:
...
    /images:
        post:
            summary: "Upload Image"
            description: "Uploads a new image for the authenticated user."
            tags: ["Images"]
            security:
                - cookieAuth: [] # Requires auth
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            properties:
                                file:
                                    type: string
                                    format: binary
                                    description: "The image file to upload (JPEG, PNG, GIF, WEBP)."
                            required:
                                - file
            responses:
                "201":
                    description: "Image uploaded successfully"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ImageMetadata"
                "400":
                    description: "Bad Request (e.g., missing file, invalid type, size limit)"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "401":
                    description: "Unauthorized"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "500":
                    description: "Internal Server Error (e.g., storage failed, DB failed)"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
        get:
            summary: "List User Images"
            description: "Retrieves metadata for all images uploaded by the authenticated user."
            tags: ["Images"]
            security:
                - cookieAuth: [] # Requires auth
            responses:
                "200":
                    description: "List of image metadata"
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/ImageMetadata"
                "401":
                    description: "Unauthorized"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "500":
                    description: "Internal Server Error"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
...
components:
    schemas:
        ImageMetadata:
            type: object
            properties:
                id:
                    type: string
                    format: uuid
                    description: "Unique identifier for the image."
                user_id:
                    type: integer
                    format: int64
                    description: "ID of the user who uploaded the image."
                filename:
                    type: string
                    description: "Original filename of the uploaded image."
                content_type:
                    type: string
                    description: "MIME type of the image (e.g., 'image/jpeg')."
                size:
                    type: integer
                    format: int64
                    description: "Size of the image in bytes."
                width:
                    type: integer
                    description: "Width of the image in pixels (if available)."
                height:
                    type: integer
                    description: "Height of the image in pixels (if available)."
                created_at:
                    type: string
                    format: date-time
                    description: "Timestamp when the image metadata was created."
                updated_at:
                    type: string
                    format: date-time
                    description: "Timestamp when the image metadata was last updated."
            required:
                - id
                - user_id
                - filename
                - content_type
                - size
                - created_at
                - updated_at
...
        ErrorResponse:
            type: object
            properties:
                error:
                    type: string
                    description: "A description of the error."
            required:
                - error
...
    securitySchemes:
        cookieAuth: # Define cookie-based authentication
            type: apiKey
            in: cookie
            name: auth_token # Name of the cookie